fix 5 bugs, cache shard_wal, tighten type safety
- fix delayed_ct stats truncated to 32 bits (%u cast to uint → PRIu64)
- fix pause-time-left uint64 wraparound on negative delta (uint64 → int64 + clamp)
- fix WAL stats (nmig, nrec) showing zero when sharding active (aggregate across shards)
- fix EMFILE/ENFILE causing 100% CPU busy-loop (deregister listen socket from epoll,
  re-register in connclose when fd freed; sockwant DEL now resets added flag on linux/darwin)
- fix scan_line_end missing \r\n after bare \r (loop instead of single memchr)
- cache shard_wal result in enqueue_job, bury_job, OP_DELETE (1 hash per op, was 2–3)
- replace epollq_rmconn rebuild-and-reverse with pointer-to-pointer scan
- change fmt_job_stats and fmt_stats_tube to take void* (removes UB function-pointer casts)
- change zalloc(int) to zalloc(size_t) for correct allocation type
- update README: bug count 32 → 35, document EMFILE backpressure, stats fixes, internal opts

+++ dat.h
@@ -292,7 +292,7 @@ struct Tube {
 void warn(const char *fmt, ...) __attribute__((format(printf, 1, 2)));
 void warnx(const char *fmt, ...) __attribute__((format(printf, 1, 2)));
 char* fmtalloc(char *fmt, ...) __attribute__((format(printf, 1, 2)));
-void* zalloc(int n);
+void* zalloc(size_t n);
 #define new(T) zalloc(sizeof(T))
 void optparse(Server*, char**);
